Using terminal callbacks to put on statusline the test status.
Using terminal callbacks, now it's possible to parse the test output and
analyses whether the test passed or not.  To get a clear feedback
neoterm is using 3 possible states for a test: 'RUNNING', 'SUCCESS',
'FAILED', for each one we have a new highlight group to be used in the
statusline.