Add support to many terminal buffers
From now on, it'll be possible to manage many terminal buffers with
neoterm. Besides the usual commands, `:T`, `:Tclose`, etc, you now have
`:Tnew` which will open a new terminal split. Each terminal split has an
id which is incremental, from 1, and is on the end of the buffer name
(like `;#neoterm-1`).

The usual commands will still working, always sending the commands to
the last opened terminal, but you'll be able to send commands to a
specific terminal with the terminal id as suffix of the command, like:
`:T2 ls` which will send an `ls` to terminal 2.