Select REPL by filetype not file extension
This leaves filetype identification to dedicated plugins, and thus
avoids code duplication. It also means the user can override the
selected REPL by simply setting the filetype.

However, this breaks Julia and Idris REPL selection in a default install
(`.jl` files are identified as `filetype=lisp` and `.[l]idr` files are
not identified at all.

This is easy to solve though: the user can either add autocommands to
his/her vimrc to set the filetype, or use a plugin like julia-vim to do
it for them. (There is a note on this in the README.)