Split TREPLSend to TREPLSendLine and TREPLSendSelection
neoterm#repl#selection tries to detect visual mode by manipulating the
bookmarks '< and '>, however, it breaks the usage of `gv`. Instead, the right
way to handle the visual mode and normal mode is to dispatch different
functions for different modes.

It is unavoidable to change the public API: the original commands `TREPLSend`
is now split to two different commands `TREPLSendLine` and
`TREPLSendSelection`.