Remove test feature (#126)
With the intention to keep this plugin easier to maintain, and also
avoiding too many vim/nvim plugins with the same goals, the test feature
(deprecated in December of 2016) is being removed.

This was decided after a short talk with the [vim-test](https://github.com/janko-m/vim-test) creator.

The vim-test already have a lot of test logic, such as: choosing
commands, and strutures of how run the tests. So, doesn't make too much
sense to keep all this logic on neoterm too.

I know that vim-test does not have all the features that neoterm#test
feature use to have, but we can contribute with vim-test to make it keep
the missing features.