Execute `G` when we reopen the terminal.
So a little motivation, about why I think this change should be done:
I've set in my vimrc  `autocmd TermOpen * set bufhidden=hide` which
hides the terminal instead of closing it when using `:q`, but an
unexpected side effect was that when I execute a command while the
buffer is hidden, it would not execute the scrolldown.

That's why I hooked the scrolldown function on the reopen, so that when
I use :Topen I have the buffer scrolled down.

How I did it? Just copied the same hook in the command execution. I can
change it to a function to mitigate copying this around if you request?