Fix `:Tclose` and `:TcloseAll`
When the default behavior changed to not replace the current buffer
these commands stop working. To fix, besides saving the window of
orignin, not the buffer of origin is also being saved, when closing a
neoterm buffer if there's no window to return, neoterm will return to
the origin buffer.

Fixes #271.